Create Research into the problem of payment system failures
The project demonstrates the main skills of working in Microsoft Excel. "As part of the analytical project, a study of the problem of failures in the operation of payment systems was conducted. The data was analysed using Microsoft Excel. The main causes of failures were identified, recommendations for their elimination were developed and measures to optimise the operation of payment systems were proposed".

Tools: MS Excel, Power Query, Power Pivot, Google Sheets

Field of activity: Analytics in the retail sector

The main activity: Data Analysis, Business Analysis